Refactor common code for code auth flow to AuthenticationProvider
