+++ src/OneDriveSdk/Authentication/AuthenticationProvider.cs
@@ -22,7 +22,9 @@
 
 namespace Microsoft.OneDrive.Sdk
 {
+    using System;
     using System.Collections.Generic;
+    using System.Net;
     using System.Net.Http;
     using System.Net.Http.Headers;
     using System.Text;
@@ -146,6 +148,66 @@ namespace Microsoft.OneDrive.Sdk
             }
 
             return null;
+        }
+
+        internal async Task<string> GetAuthorizationCodeAsync(string returnUrl = null)
+        {
+            if (this.ServiceInfo.WebAuthenticationUi != null)
+            {
+                returnUrl = returnUrl ?? this.ServiceInfo.ReturnUrl;
+
+                var requestUriStringBuilder = new StringBuilder();
+                requestUriStringBuilder.Append(this.ServiceInfo.AuthenticationServiceUrl);
+                requestUriStringBuilder.AppendFormat("?{0}={1}", Constants.Authentication.RedirectUriKeyName, returnUrl);
+                requestUriStringBuilder.AppendFormat("&{0}={1}", Constants.Authentication.ClientIdKeyName, this.ServiceInfo.AppId);
+                requestUriStringBuilder.AppendFormat("&{0}={1}", Constants.Authentication.ScopeKeyName, string.Join("%20", this.ServiceInfo.Scopes));
+
+                if (!string.IsNullOrEmpty(this.ServiceInfo.UserId))
+                {
+                    requestUriStringBuilder.AppendFormat("&{0}={1}", Constants.Authentication.UserIdKeyName, this.ServiceInfo.UserId);
+                }
+
+                requestUriStringBuilder.AppendFormat("&{0}={1}", Constants.Authentication.ResponseTypeKeyName, Constants.Authentication.CodeKeyName);
+
+                var requestUri = new Uri(requestUriStringBuilder.ToString());
+
+                var authenticationResponseValues = await this.ServiceInfo.WebAuthenticationUi.AuthenticateAsync(
+                    requestUri,
+                    new Uri(returnUrl));
+                OAuthErrorHandler.ThrowIfError(authenticationResponseValues);
+
+                string code;
+                if (authenticationResponseValues != null && authenticationResponseValues.TryGetValue("code", out code))
+                {
+                    return code;
+                }
+            }
+
+            return null;
+        }
+
+        internal string GetCodeRedemptionRequestBody(string code, string returnUrl = null)
+        {
+            returnUrl = returnUrl ?? this.ServiceInfo.ReturnUrl;
+
+            var requestBodyString = string.Format(
+                "{0}={1}&{2}={3}&{4}={5}&{6}={7}&{8}=authorization_code",
+                Constants.Authentication.RedirectUriKeyName,
+                returnUrl,
+                Constants.Authentication.ClientIdKeyName,
+                this.ServiceInfo.AppId,
+                Constants.Authentication.ScopeKeyName,
+                WebUtility.UrlEncode(string.Join(" ", this.ServiceInfo.Scopes)),
+                Constants.Authentication.CodeKeyName,
+                code,
+                Constants.Authentication.GrantTypeKeyName);
+
+            if (!string.IsNullOrEmpty(this.ServiceInfo.ClientSecret))
+            {
+                requestBodyString += "&client_secret=" + this.ServiceInfo.ClientSecret;
+            }
+
+            return requestBodyString;
         }
 
         protected virtual Task<AccountSession> RefreshAccessTokenAsync(string refreshToken)
